Move RLP encode / decode methods into a test util
These methods should never be used in production.
